# services/asp/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Untch — spend control for agents",
    "version": "1.0.0",
    "description": "Every service this host offers, generated from the definitions its own validators enforce. Each operation carries what must already exist before it can succeed, what it will change, whether repeating it is safe, and what it costs."
  },
  "servers": [
    {
      "url": "https://asp.untch.xyz"
    }
  ],
...
  "paths": {
...
    "/consumer/account/link/start": {
      "post": {
        "operationId": "account_link_start",
        "summary": "Opens a one-time request to bind a wallet — and optionally a marketplace identity — to an Untch account.",
        "description": "Opens a one-time request to bind a wallet — and optionally a marketplace identity — to an Untch account. For a marketplace caller whose agent id Untch has never seen, or a person signing in for the first time.\n\nYou provide: Any one of: requestedScopes, marketplace, marketplaceAgentId, marketplaceBuyerId, taskRef, serviceOrderRef, returnUrl.\n\nYou receive: a link request id, a one-time code shown exactly once, the message the wallet must sign, and the URL to sign it at.",
        "tags": [
          "A2MCP"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "AccountLinkStart",
                "properties": {
                  "requestedScopes": {
                    "type": "array",
                    "description": "identity proves who you are. policy-authority additionally permits owning a policy.",
                    "items": {
                      "type": "string",
                      "enum": [
                        "identity",
                        "policy-authority"
                      ]
                    }
                  },
                  "marketplace": {
                    "type": "string",
                    "description": "The marketplace this call arrived through, e.g. okx."
                  },
                  "marketplaceAgentId": {
                    "type": "string",
                    "description": "UNPROVEN on arrival. Recorded as context and authorises nothing until a wallet signs for it."
                  },
                  "marketplaceBuyerId": {
                    "type": "string"
                  },
                  "taskRef": {
                    "type": "string",
                    "description": "The marketplace task or job this call belongs to."
                  },
                  "serviceOrderRef": {
                    "type": "string"
                  },
                  "returnUrl": {
                    "type": "string",
                    "description": "Where to return afterwards. Matched against an allowlist by exact ORIGIN, and must be https."
                  }
                },
                "additionalProperties": false
              },
              "examples": {
                "valid": {
                  "summary": "Link a marketplace caller",
                  "value": {
                    "marketplace": "okx",
                    "marketplaceAgentId": "6047",
                    "taskRef": "task-42",
                    "requestedScopes": [
                      "identity"
                    ]
                  }
                },
                "refused": {
                  "summary": "Ask for a scope that does not exist → UNKNOWN_SCOPE",
                  "value": {
                    "requestedScopes": [
                      "spend-anything"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "a link request id, a one-time code shown exactly once, the message the wallet must sign, and the URL to sign it at",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "AccountLinkStarted",
                  "properties": {
                    "linkRequestId": {
                      "type": "string",
                      "pattern": "^ulnk_[a-z0-9]{26}$"
                    },
                    "oneTimeCode": {
                      "type": "string",
                      "description": "Returned once. Stored hashed; no later read can produce it."
                    },
                    "expiresAt": {
                      "type": "string",
                      "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d+)?Z$",
                      "description": "When this link request stops being redeemable.",
                      "examples": [
                        "2026-08-01T12:00:00.000Z"
                      ]
                    },
                    "proofMethod": {
                      "type": "string",
                      "const": "siwe-personal-sign"
                    },
                    "walletActionUrl": {
                      "type": "string"
                    },
                    "requestedScopes": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "linkRequestId",
                    "oneTimeCode",
                    "expiresAt",
                    "proofMethod"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "UNKNOWN_SCOPE — requestedScopes contains a scope this host does not grant; RETURN_URL_NOT_ALLOWED — returnUrl is not an exact-origin match for an allowed origin, or is not https",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Refusal",
                  "description": "The shape of every refusal this service returns, at every status code.",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "A stable machine-readable reason. Safe to branch on."
                    },
                    "message": {
                      "type": "string",
                      "description": "One sentence a caller can act on."
                    },
                    "retryable": {
                      "type": "boolean",
                      "description": "Whether repeating the identical request could succeed without changing it."
                    },
                    "docsUrl": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "description": "Where the rule is written down, when it is."
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Present only on an unclassified internal failure. Quote it when reporting; it names the log line holding the cause."
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retryable"
                  ]
                }
              }
            }
          },
          "503": {
            "description": "ACCOUNT_LINK_UNAVAILABLE — this instance cannot mint sessions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Refusal",
                  "description": "The shape of every refusal this service returns, at every status code.",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "A stable machine-readable reason. Safe to branch on."
                    },
                    "message": {
                      "type": "string",
                      "description": "One sentence a caller can act on."
                    },
                    "retryable": {
                      "type": "boolean",
                      "description": "Whether repeating the identical request could succeed without changing it."
                    },
                    "docsUrl": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "description": "Where the rule is written down, when it is."
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Present only on an unclassified internal failure. Quote it when reporting; it names the log line holding the cause."
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retryable"
                  ]
                }
              }
            }
          }
        },
        "x-untch-pricing": {
          "kind": "free",
          "price": null,
          "amountBaseUnits": null
        },
        "x-untch-maturity": "live",
        "x-untch-class": "ACCOUNT_CONTROL",
        "x-untch-predecessors": [],
        "x-untch-side-effects": [
          {
            "what": "Creates a pending link request holding a hashed one-time code.",
            "durable": true
          },
          {
            "what": "Approves no payment. No route reachable from this code takes an amount.",
            "durable": false
          }
        ],
        "x-untch-idempotency": "not-idempotent",
        "x-untch-schema-version": "1.0.0"
      }
    },
...
  },
...
}